add support check

+++ index.js
@@ -1,5 +1,20 @@
 class Storage {
   /**
+   * 构造器
+   * constructor
+   */
+  constructor () {
+    this._isSupport = true
+    try {
+      const test = 'localStorage support test'
+      window.localStorage.setItem('test', test)
+      window.localStorage.removeItem('test')
+    } catch (e) {
+      this._isSupport = false
+    }
+  }
+
+  /**
    * 获取时间戳
    */
   static get timestamp() {
@@ -21,6 +36,10 @@ class Storage {
    * @param {Number} expired_second 过期时间 单位秒
    */
   set (key, value, expired_second) {
+    if (!this._isSupport) {
+      return null
+    }
+
     // 存储实例
     const entity = {
       timestamp: Storage.timestamp,
@@ -37,6 +56,10 @@ class Storage {
    * @param {String} key 键
    */
   get (key) {
+    if (!this._isSupport) {
+      return null
+    }
+
     let entity
     try {
       entity = window.localStorage.getItem(key)
@@ -66,6 +89,10 @@ class Storage {
    * @param {String} key 键
    */
   remove (key) {
+    if (!this._isSupport) {
+      return null
+    }
+
     try {
       window.localStorage.removeItem(key)
     } catch (err) {
@@ -78,6 +105,10 @@ class Storage {
    * 清空存储
    */
   clear () {
+    if (!this._isSupport) {
+      return null
+    }
+
     try {
       window.localStorage.clear()
     } catch (err) {
